Unify the github and native knolwedge wizards
