+++ src/components/Contribute/Knowledge/AttributionInformation/AttributionInformation.tsx
@@ -1,31 +1,25 @@
 import React, { useEffect } from 'react';
-import { checkKnowledgeFormCompletion } from '../validation';
 import { KnowledgeFormData } from '@/types';
 import { ValidatedOptions, FormFieldGroupHeader, FormGroup, TextInput, FormHelperText, HelperText, HelperTextItem } from '@patternfly/react-core';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 
 interface Props {
-  reset: boolean;
   isEditForm?: boolean;
   knowledgeFormData: KnowledgeFormData;
-  setDisableAction: React.Dispatch<React.SetStateAction<boolean>>;
   titleWork: string;
-  setTitleWork: React.Dispatch<React.SetStateAction<string>>;
+  setTitleWork: (val: string) => void;
   linkWork: string;
-  setLinkWork: React.Dispatch<React.SetStateAction<string>>;
+  setLinkWork: (val: string) => void;
   revision: string;
-  setRevision: React.Dispatch<React.SetStateAction<string>>;
+  setRevision: (val: string) => void;
   licenseWork: string;
-  setLicenseWork: React.Dispatch<React.SetStateAction<string>>;
+  setLicenseWork: (val: string) => void;
   creators: string;
-  setCreators: React.Dispatch<React.SetStateAction<string>>;
+  setCreators: (val: string) => void;
 }
 
 const AttributionInformation: React.FC<Props> = ({
-  reset,
   isEditForm,
-  knowledgeFormData,
-  setDisableAction,
   titleWork,
   setTitleWork,
   linkWork,
@@ -44,14 +38,6 @@ const AttributionInformation: React.FC<Props> = ({
   const [validCreators, setValidCreators] = React.useState<ValidatedOptions>();
 
   useEffect(() => {
-    setValidTitle(ValidatedOptions.default);
-    setValidLink(ValidatedOptions.default);
-    setValidRevision(ValidatedOptions.default);
-    setValidLicense(ValidatedOptions.default);
-    setValidCreators(ValidatedOptions.default);
-  }, [reset]);
-
-  useEffect(() => {
     if (!isEditForm) {
       return;
     }
@@ -66,10 +52,8 @@ const AttributionInformation: React.FC<Props> = ({
     const title = titleStr.trim();
     if (title.length > 0) {
       setValidTitle(ValidatedOptions.success);
-      setDisableAction(!checkKnowledgeFormCompletion(knowledgeFormData));
       return;
     }
-    setDisableAction(true);
     setValidTitle(ValidatedOptions.error);
     return;
   };
@@ -77,56 +61,30 @@ const AttributionInformation: React.FC<Props> = ({
   const validateLink = (linkStr: string) => {
     const link = linkStr.trim();
     if (link.length === 0) {
-      setDisableAction(true);
       setValidLink(ValidatedOptions.error);
       return;
     }
     try {
       new URL(link);
       setValidLink(ValidatedOptions.success);
-      setDisableAction(!checkKnowledgeFormCompletion(knowledgeFormData));
-      return;
     } catch (e) {
-      setDisableAction(true);
       setValidLink(ValidatedOptions.warning);
-      return;
     }
   };
 
   const validateRevision = (revisionStr: string) => {
     const revision = revisionStr.trim();
-    if (revision.length > 0) {
-      setValidRevision(ValidatedOptions.success);
-      setDisableAction(!checkKnowledgeFormCompletion(knowledgeFormData));
-      return;
-    }
-    setDisableAction(true);
-    setValidRevision(ValidatedOptions.error);
-    return;
+    setValidRevision(revision.length > 0 ? ValidatedOptions.success : ValidatedOptions.error);
   };
 
   const validateLicense = (licenseStr: string) => {
     const license = licenseStr.trim();
-    if (license.length > 0) {
-      setValidLicense(ValidatedOptions.success);
-      setDisableAction(!checkKnowledgeFormCompletion(knowledgeFormData));
-      return;
-    }
-    setDisableAction(true);
-    setValidLicense(ValidatedOptions.error);
-    return;
+    setValidLicense(license.length > 0 ? ValidatedOptions.success : ValidatedOptions.error);
   };
 
   const validateCreators = (creatorsStr: string) => {
     const creators = creatorsStr.trim();
-    if (creators.length > 0) {
-      setValidCreators(ValidatedOptions.success);
-      setDisableAction(!checkKnowledgeFormCompletion(knowledgeFormData));
-      return;
-    }
-    setDisableAction(true);
-    setValidCreators(ValidatedOptions.error);
-    return;
+    setValidCreators(creators.length > 0 ? ValidatedOptions.success : ValidatedOptions.error);
   };
 
   return (
