+++ src/components/Contribute/submitUtils.ts
@@ -1,7 +1,11 @@
-import { KnowledgeFormData, KnowledgeYamlData } from '@/types';
+import { AttributionData, KnowledgeEditFormData, KnowledgeFormData, KnowledgeYamlData } from '@/types';
 import { KnowledgeSchemaVersion } from '@/types/const';
 import { dumpYaml } from '@/utils/yamlConfig';
 import { ActionGroupAlertContent } from '@/components/Contribute/types';
+import { amendCommit, getGitHubUsername, updatePullRequest } from '@/utils/github';
+import { Session } from 'next-auth';
+
+const KNOWLEDGE_DIR = 'knowledge/';
 
 export const submitNativeKnowledgeData = async (
   knowledgeFormData: KnowledgeFormData,
@@ -80,4 +84,300 @@ export const submitNativeKnowledgeData = async (
   };
   setActionGroupAlertContent(actionGroupAlertContent);
   return true;
+};
+
+export const submitGithubKnowledgeData = async (
+  knowledgeFormData: KnowledgeFormData,
+  githubUsername: string,
+  setActionGroupAlertContent: (content: ActionGroupAlertContent) => void
+): Promise<boolean> => {
+  // Strip leading slash and ensure trailing slash in the file path
+  let sanitizedFilePath = knowledgeFormData.filePath!.startsWith('/') ? knowledgeFormData.filePath!.slice(1) : knowledgeFormData.filePath;
+  sanitizedFilePath = sanitizedFilePath!.endsWith('/') ? sanitizedFilePath : `${sanitizedFilePath}/`;
+
+  const knowledgeYamlData: KnowledgeYamlData = {
+    created_by: githubUsername!,
+    version: KnowledgeSchemaVersion,
+    domain: knowledgeFormData.domain!,
+    document_outline: knowledgeFormData.documentOutline!,
+    seed_examples: knowledgeFormData.seedExamples.map((example) => ({
+      context: example.context,
+      questions_and_answers: example.questionAndAnswers.map((questionAndAnswer) => ({
+        question: questionAndAnswer.question,
+        answer: questionAndAnswer.answer
+      }))
+    })),
+    document: {
+      repo: knowledgeFormData.knowledgeDocumentRepositoryUrl!,
+      commit: knowledgeFormData.knowledgeDocumentCommit!,
+      patterns: knowledgeFormData.documentName!.split(',').map((pattern) => pattern.trim())
+    }
+  };
+
+  const yamlString = dumpYaml(knowledgeYamlData);
+
+  const attributionData: AttributionData = {
+    title_of_work: knowledgeFormData.titleWork!,
+    link_to_work: knowledgeFormData.linkWork!,
+    revision: knowledgeFormData.revision!,
+    license_of_the_work: knowledgeFormData.licenseWork!,
+    creator_names: knowledgeFormData.creators!
+  };
+
+  const waitForSubmissionAlert: ActionGroupAlertContent = {
+    title: 'Knowledge contribution submission in progress.!',
+    message: `Once the submission is successful, it will provide the link to the newly created Pull Request.`,
+    success: true,
+    waitAlert: true,
+    timeout: false
+  };
+  setActionGroupAlertContent(waitForSubmissionAlert);
+
+  const { name, submissionSummary, documentOutline, email } = knowledgeFormData;
+  const response = await fetch('/api/github/pr/knowledge', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify({
+      content: yamlString,
+      attribution: attributionData,
+      name,
+      email,
+      submissionSummary,
+      documentOutline,
+      filePath: sanitizedFilePath
+    })
+  });
+
+  if (!response.ok) {
+    const actionGroupAlertContent: ActionGroupAlertContent = {
+      title: `Failed data submission`,
+      message: response.statusText,
+      success: false
+    };
+    setActionGroupAlertContent(actionGroupAlertContent);
+    return false;
+  }
+
+  const result = await response.json();
+  const actionGroupAlertContent: ActionGroupAlertContent = {
+    title: 'Knowledge contribution submitted successfully!',
+    message: `Thank you for your contribution!`,
+    url: `${result.html_url}`,
+    success: true
+  };
+  setActionGroupAlertContent(actionGroupAlertContent);
+  return true;
+};
+
+export const updateNativeKnowledgeData = async (
+  knowledgeFormData: KnowledgeFormData,
+  knowledgeEditFormData: KnowledgeEditFormData,
+  setActionGroupAlertContent: (content: ActionGroupAlertContent) => void
+): Promise<boolean> => {
+  // Strip leading slash and ensure trailing slash in the file path
+  let sanitizedFilePath = knowledgeFormData.filePath!.startsWith('/') ? knowledgeFormData.filePath!.slice(1) : knowledgeFormData.filePath;
+  sanitizedFilePath = sanitizedFilePath!.endsWith('/') ? sanitizedFilePath : `${sanitizedFilePath}/`;
+
+  const knowledgeYamlData: KnowledgeYamlData = {
+    created_by: knowledgeFormData.email,
+    version: KnowledgeSchemaVersion,
+    domain: knowledgeFormData.domain!,
+    document_outline: knowledgeFormData.documentOutline!,
+    seed_examples: knowledgeFormData.seedExamples.map((example) => ({
+      context: example.context,
+      questions_and_answers: example.questionAndAnswers.map((questionAndAnswer) => ({
+        question: questionAndAnswer.question,
+        answer: questionAndAnswer.answer
+      }))
+    })),
+    document: {
+      repo: knowledgeFormData.knowledgeDocumentRepositoryUrl!,
+      commit: knowledgeFormData.knowledgeDocumentCommit!,
+      patterns: knowledgeFormData.documentName!.split(',').map((pattern) => pattern.trim())
+    }
+  };
+
+  const yamlString = dumpYaml(knowledgeYamlData);
+
+  const waitForSubmissionAlert: ActionGroupAlertContent = {
+    title: 'Knowledge contribution submission in progress!',
+    message: `Once the submission is successful, it will provide the link to the newly created Pull Request.`,
+    success: true,
+    waitAlert: true,
+    timeout: false
+  };
+  setActionGroupAlertContent(waitForSubmissionAlert);
+
+  const { name, email, submissionSummary, documentOutline } = knowledgeFormData;
+  const { branchName, oldFilesPath } = knowledgeEditFormData;
+  const response = await fetch('/api/native/pr/knowledge', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify({
+      action: 'update',
+      branchName: branchName,
+      content: yamlString,
+      name,
+      email,
+      submissionSummary,
+      documentOutline,
+      filePath: sanitizedFilePath,
+      oldFilesPath: oldFilesPath
+    })
+  });
+
+  if (!response.ok) {
+    const actionGroupAlertContent: ActionGroupAlertContent = {
+      title: `Failed data submission`,
+      message: response.statusText,
+      success: false
+    };
+    setActionGroupAlertContent(actionGroupAlertContent);
+    return false;
+  }
+
+  await response.json();
+  const actionGroupAlertContent: ActionGroupAlertContent = {
+    title: 'Knowledge contribution updated successfully!',
+    message: `Thank you for your contribution!`,
+    url: '/dashboard/',
+    success: true
+  };
+  setActionGroupAlertContent(actionGroupAlertContent);
+  return true;
+};
+
+export const updateGithubKnowledgeData = async (
+  session: Session | null,
+  knowledgeFormData: KnowledgeFormData,
+  knowledgeEditFormData: KnowledgeEditFormData,
+  setActionGroupAlertContent: (content: ActionGroupAlertContent) => void
+): Promise<boolean> => {
+  if (session?.accessToken) {
+    try {
+      console.log(`Updating PR with number: ${knowledgeEditFormData.pullRequestNumber}`);
+      await updatePullRequest(session.accessToken, knowledgeEditFormData.pullRequestNumber, {
+        title: knowledgeFormData.submissionSummary,
+        body: knowledgeFormData.documentOutline
+      });
+
+      const headers = {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${session.accessToken}`,
+        Accept: 'application/vnd.github+json',
+        'X-GitHub-Api-Version': '2022-11-28'
+      };
+
+      const githubUsername = await getGitHubUsername(headers);
+      console.log(`GitHub username: ${githubUsername}`);
+
+      const knowledgeYamlData: KnowledgeYamlData = {
+        created_by: githubUsername!,
+        version: KnowledgeSchemaVersion,
+        domain: knowledgeFormData.domain!,
+        document_outline: knowledgeFormData.documentOutline!,
+        seed_examples: knowledgeFormData.seedExamples.map((example) => ({
+          context: example.context,
+          questions_and_answers: example.questionAndAnswers.map((questionAndAnswer) => ({
+            question: questionAndAnswer.question,
+            answer: questionAndAnswer.answer
+          }))
+        })),
+        document: {
+          repo: knowledgeFormData.knowledgeDocumentRepositoryUrl!,
+          commit: knowledgeFormData.knowledgeDocumentCommit!,
+          patterns: knowledgeFormData.documentName!.split(',').map((pattern) => pattern.trim())
+        }
+      };
+
+      const yamlString = dumpYaml(knowledgeYamlData);
+      console.log('Updated knowledge YAML content:', yamlString);
+
+      const attributionData: AttributionData = {
+        title_of_work: knowledgeFormData.titleWork!,
+        link_to_work: knowledgeFormData.linkWork!,
+        revision: knowledgeFormData.revision!,
+        license_of_the_work: knowledgeFormData.licenseWork!,
+        creator_names: knowledgeFormData.creators!
+      };
+      const attributionContent = `Title of work: ${attributionData.title_of_work}
+Link to work: ${attributionData.link_to_work}
+Revision: ${attributionData.revision}
+License of the work: ${attributionData.license_of_the_work}
+Creator names: ${attributionData.creator_names}
+`;
+
+      console.log('Updated knowledge attribution content:', attributionData);
+
+      const commitMessage = `Amend commit with updated content\n\nSigned-off-by: ${knowledgeFormData.name} <${knowledgeFormData.email}>`;
+
+      // Ensure proper file paths for the edit
+      const finalYamlPath = KNOWLEDGE_DIR + knowledgeFormData.filePath.replace(/^\//, '').replace(/\/?$/, '/') + 'qna.yaml';
+      const finalAttributionPath = KNOWLEDGE_DIR + knowledgeFormData.filePath.replace(/^\//, '').replace(/\/?$/, '/') + 'attribution.txt';
+
+      const oldFilePath = {
+        yaml: KNOWLEDGE_DIR + knowledgeEditFormData.oldFilesPath.replace(/^\//, '').replace(/\/?$/, '/') + 'qna.yaml',
+        attribution: KNOWLEDGE_DIR + knowledgeEditFormData.oldFilesPath.replace(/^\//, '').replace(/\/?$/, '/') + 'attribution.txt'
+      };
+
+      console.log('Knowledge update old file path : ', oldFilePath);
+
+      const newFilePath = {
+        yaml: finalYamlPath,
+        attribution: finalAttributionPath
+      };
+      console.log('Knowledge update new file path : ', oldFilePath);
+
+      const waitForSubmissionAlert: ActionGroupAlertContent = {
+        title: 'Knowledge contribution update is in progress.!',
+        message: `Once the update is successful, it will provide the link to the updated Pull Request.`,
+        success: true,
+        waitAlert: true,
+        timeout: false
+      };
+      setActionGroupAlertContent(waitForSubmissionAlert);
+
+      const res = await fetch('/api/envConfig');
+      const envConfig = await res.json();
+
+      const amendedCommitResponse = await amendCommit(
+        session.accessToken,
+        githubUsername,
+        envConfig.UPSTREAM_REPO_NAME,
+        oldFilePath,
+        newFilePath,
+        yamlString,
+        attributionContent,
+        knowledgeEditFormData.branchName,
+        commitMessage
+      );
+      console.log('Amended commit response:', amendedCommitResponse);
+
+      const prLink = `https://github.com/${envConfig.UPSTREAM_REPO_OWNER}/${envConfig.UPSTREAM_REPO_NAME}/pull/${knowledgeEditFormData.pullRequestNumber}`;
+      const actionGroupAlertContent: ActionGroupAlertContent = {
+        title: 'Knowledge contribution updated successfully!',
+        message: `Thank you for your contribution!`,
+        url: `${prLink}`,
+        success: true
+      };
+      setActionGroupAlertContent(actionGroupAlertContent);
+      // Knowledge is updated, wait for a bit and let's go back to dashboard.
+      await new Promise((r) => setTimeout(r, 4000));
+    } catch (error) {
+      console.error('Error updating PR:', error);
+      const actionGroupAlertContent: ActionGroupAlertContent = {
+        title: `Failed to update PR with number: ${knowledgeEditFormData.pullRequestNumber}`,
+        message: `PR update failed because of ${error}`,
+        success: false
+      };
+      setActionGroupAlertContent(actionGroupAlertContent);
+      return false;
+    }
+    return true;
+  }
+  return false;
 };
